Add files via upload
A clock is a device used to tell time. Moving hands on the face of a clock point to the current hour, minute, and second. A clock can be big enough to be in a giant tower. A clock can also be small enough to fit on a person's wrist, where it is called a watch.